finished step 3
