+++ README.md
@@ -1,1 +1,4 @@
 # http-server
+Here is our simple http server.
+It can use the client request to find and show what is being requested.
+I used stack overflow to find out about the os.path.isfile method to validate whether the uri is infact a valid file.
